Move requestDomains to domains.requests
